requirements.txt
h5py==3.10.0
matplotlib==3.7.1
numpy==1.24.2
opencv_python==4.8.0.74
pandas==2.0.3
Pillow==10.2.0
scikit_learn==1.3.0
tensorboardX==2.6.2.2
torch==2.0.0+cu118
torchvision==0.15.1+cu118
tqdm==4.61.2
transformers==4.38.2
